requirements.txt

Flask
python-dotenv
gunicorn
flask-cors
